+++ test/signing/test_sign.py
@@ -1,8 +1,7 @@
+import pytest
+import sys
+
 from pathlib import Path
-from sys import version_info
-
-
-from pytest import mark
 from fpdf import FPDF
 from test.conftest import assert_pdf_equal, check_signature, EPOCH
 
@@ -20,6 +19,10 @@ def test_sign_pkcs12(tmp_path):
     check_signature(pdf, TRUSTED_CERT_PEMS)
 
 
+@pytest.mark.skipif(
+    sys.platform in ("cygwin", "win32") and sys.version_info[:2] == (3, 14),
+    reason="Skipped on Windows with Python 3.14 due to zlib compressed data differences",
+)
 def test_sign_pkcs12_with_link(tmp_path):
     "This test ensures that Signature & Link annotations can be combined"
     pdf = FPDF()